Create SummaryOfIntent.txt
This is just a statement of intent with this project.  Hopefully, I will find time to continue working on it.  I picture myself lugging a laptop to play pool.  I guess this message doesn't belong in the commit message, but I'm just so excited to be here writing this message.  This represents an important milestone in this project: the first file.  But, it's not just the first file, it's the project statement.  It's like that famous thing the moon-walker said when he took that first step: one small stoop for man, one gigantic leap to ..., I don't remember.  But, you get the point. Ok, well, I'm at work so I'd better get back to it.  I don't think they would want me doing this instead of watching the security cameras.  The DOD is funny about those things.  

So, in short, this is my first commit.